Add sumiru + Infuse medium
This adds:
sumiru square panel (92900266401)
Infuse medium ceiling light (915005997201)
And fixes the title formated errors in the log.
Unfortunatly as the flows are not filtered on a driver, but the class, it will need the [[device]] in the title formated object, they don't show up in the flow card itself though, just the example.